fix: apply PR review feedback for encode_pre_encoded API
- Accept PyAny (not PyBytes) in StreamingEncoder.write_object_pre_encoded
- Force little-endian in Python cross-language golden test helper
- Explicit LE serialization in C++ cross-language helper (portable)
- Clarify verbatim byte-order semantics in docs
- Fix -l tensogram_ffi -> -ltensogram_ffi in C++ example
- Set both DYLD_LIBRARY_PATH and LD_LIBRARY_PATH in Rust cross-language test

Agent-Logs-Url: https://github.com/ecmwf/tensogram/sessions/c6c85860-39e4-4f1c-bd47-80b668415627

+++ tests/python/cross_language_pre_encoded_helper.py
@@ -17,7 +17,7 @@ import tensogram
 def main() -> None:
     # Same deterministic input as the Rust driver.
     values = np.array([200.0 + i * 0.125 for i in range(1024)], dtype=np.float64)
-    raw_bytes = values.tobytes()  # little-endian on ARM Mac
+    raw_bytes = values.astype("<f8", copy=False).tobytes()  # explicit little-endian
 
     global_meta = {"version": 2}
     descriptor = {
